Revert "Footer now uniform, AdminLogin is now persistent"
