+++ html/footer.html
@@ -5,7 +5,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Footer Test</title>
     <link rel="stylesheet" href="../css/footerStyle.css">
-    <script src="../js/footerScript.js" defer></script>
     <!-- Font Awesome CDN -->
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.6.0/css/all.min.css">
 </head>
@@ -46,7 +45,7 @@
         <div class="footer-bottom">
             <p class="copyright">&copy; <span id="current-year"></span> Maunatul Muslimeen Society of CA. All rights reserved.</p>
             <p class="copyright">Custom-built By BasiqCoders.</p> 
-            <button class="back-to-top" onclick="scrollToTop()">Back to Top</button>
+            <button class="back-to-top">Back to Top</button>
         </div>
     </footer>
