+++ htmlA/adminAbout.html
@@ -5,11 +5,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Admin About</title>
     <link rel="stylesheet" href="../cssA/adminAbout.css">
-    <script src="../jsA/adminAbout.js" defer></script>
+    <script type="module" src="../jsA/adminAbout.js" defer></script>
     <script src="https://code.jquery.com/jquery-3.7.1.min.js" integrity="sha256-/JqT3SQfawRcv/BIHPThkBvs0OEvtFFmqPF/lYI/Cxo=" crossorigin="anonymous"></script>
-    <script src="../js/firebase.js" type="module" defer></script>
-    <script src="https://www.gstatic.com/firebasejs/9.6.1/firebase-app.js" type = "module" defer></script>
-    <script src="https://www.gstatic.com/firebasejs/9.6.1/firebase-firestore.js" type = "module" defer></script>
     <script defer>
         $(function(){
             $('#header').load('adminHeader.html');
