+++ htmlA/adminDonate.html
@@ -7,12 +7,11 @@
 
         <link rel="stylesheet" href="../cssA/adminDonate.css">
        
-        <script src="../jsA/adminDonate.js" defer></script>
+        <script type="module" src="../jsA/adminDonate.js" defer></script>
         
-        <script src="https://code.jquery.com/jquery-3.7.1.min.js" integrity="sha256-/JqT3SQfawRcv/BIHPThkBvs0OEvtFFmqPF/lYI/Cxo=" crossorigin="anonymous"></script>
-        <script src="../js/firebase.js" type="module" defer></script>
-        <script src="https://www.gstatic.com/firebasejs/9.6.1/firebase-app.js" type = "module" defer></script>
-        <script src="https://www.gstatic.com/firebasejs/9.6.1/firebase-firestore.js" type = "module" defer></script>
+        <script src="https://code.jquery.com/jquery-3.7.1.min.js" 
+        integrity="sha256-/JqT3SQfawRcv/BIHPThkBvs0OEvtFFmqPF/lYI/Cxo=" 
+        crossorigin="anonymous">
         </script>
         <script defer>
             $(function(){
