+++ htmlA/adminEvent.html
@@ -8,12 +8,9 @@
         <!--CSS & Javascript-->
         <link rel="stylesheet" href="../cssA/adminEvent.css">
 
-        <script src="../jsA/adminEvent.js" defer></script>
+        <script type="module" src="../jsA/adminEvent.js" defer></script>
         <script src="https://code.jquery.com/jquery-3.7.1.min.js" integrity="sha256-/JqT3SQfawRcv/BIHPThkBvs0OEvtFFmqPF/lYI/Cxo=" crossorigin="anonymous"> </script>
         <!--This imports the codes of the adminHeader (including whatever is imported in the header HTML file)-->
-        <script src="../js/firebase.js" type="module" defer></script>
-        <script src="https://www.gstatic.com/firebasejs/9.6.1/firebase-app.js" type = "module" defer></script>
-        <script src="https://www.gstatic.com/firebasejs/9.6.1/firebase-firestore.js" type = "module" defer></script>
         <script defer>
             $(function(){
                 $('#header').load('adminHeader.html');
