+++ htmlA/adminHome.html
@@ -9,10 +9,7 @@
         <script src="https://code.jquery.com/jquery-3.7.1.min.js" integrity="sha256-/JqT3SQfawRcv/BIHPThkBvs0OEvtFFmqPF/lYI/Cxo=" crossorigin="anonymous"> </script>
 
         <link rel="stylesheet" href="../cssA/adminHome.css">
-        <script src="../jsA/adminHome.js"defer></script> 
-        <script src="../js/firebase.js" type="module" defer></script>
-        <script src="https://www.gstatic.com/firebasejs/9.6.1/firebase-app.js" type = "module" defer></script>
-        <script src="https://www.gstatic.com/firebasejs/9.6.1/firebase-firestore.js" type = "module" defer></script>
+        <script type="module" src="../jsA/adminHome.js"defer></script> 
 
         <script>
             $(function(){
