+++ jsA/adminAbout.js
@@ -1,6 +1,40 @@
-document.addEventListener("DOMContentLoaded", () => {
-    checkAuth();
+import { initializeApp } from "https://www.gstatic.com/firebasejs/11.0.1/firebase-app.js";
+import { getAuth, onAuthStateChanged} from 'https://www.gstatic.com/firebasejs/11.0.1/firebase-auth.js';
 
+//Hide this later
+const firebaseConfig = {
+  apiKey: "AIzaSyChNmvSjjLzXfWeGsKHebXgGq_AMUdKzHo",
+  authDomain: "project-musjid.firebaseapp.com",
+  projectId: "project-musjid",
+  storageBucket: "project-musjid.firebasestorage.app",
+  messagingSenderId: "445451894728",
+  appId: "1:445451894728:web:09bffcb1743ae1ecec4afd",
+  measurementId: "G-H5XN7NRJ6V"
+};
+
+// Initialize Firebase
+initializeApp(firebaseConfig);
+
+// Initialize Firebase Authentication and get a reference to the service
+const auth = getAuth();
+
+// Check if the user is authenticated
+onAuthStateChanged(auth, (user) => {
+  if (!user) {
+    // User is not signed in, redirect to login page
+    window.location.href = "adminLogin.html";
+    console.log("Page restricted until signed in");
+  } else {
+
+  }
+});
+// End: Redirect to login page if the user is not authenticated
+
+
+
+
+
+document.addEventListener("DOMContentLoaded", () => {
     const teamMembers = JSON.parse(localStorage.getItem("teamMembers") || "[]");
     const adminMembersBox = document.getElementById("adminMembersBox");
